track/intersect: 2 more matches (fn_80070434, fn_8007DADC)
fn_80070434: local unprototyped extern for GXSetZCompLoc suppresses
the u8 narrow so r31 is passed raw (matches target's 'mr r3, r31').

fn_8007DADC: widen param to u32 and CSE the (u8) narrow into a local
so both the stb and the compare share the same r0 value.

+++ include/track/intersect.h
@@ -97,7 +97,7 @@ void FUN_8007d0f8(void);
 void fn_8007D7EC(void);
 void FUN_8007d858(void);
 undefined4 FUN_8007d8a8(void);
-void fn_8007DADC(u8 param_1);
+void fn_8007DADC(u32 param_1);
 void fn_8007DB04(void);
 s32 fn_8007DB10(void);
 undefined4
